fix fpc compilation
git-svn-id: http://svn.code.sf.net/p/zeoslib/code-0/branches/testing-7.3@4801 d681b350-854c-439c-a81d-c0ce20710203

+++ src/core/ZVariant.pas
@@ -1036,7 +1036,7 @@ DoWideCompare:
           Result := 0;
       end;
     vtPointer:
-      Result := sign({%H-}NativeUInt(Value1.VPointer) - GetAsUInteger(Value2));
+      Result := sign(Int64({%H-}NativeUInt(Value1.VPointer) - GetAsUInteger(Value2)));
     else
       Result := 0;
   end;
